chore(lint): adopt an explicit ruff rule set and clear its findings
- ruff ran on its bare defaults, so the CI gate only checked pyflakes and a
  handful of pycodestyle errors; pyproject now names every selected rule so a
  ruff upgrade cannot silently change what the gate means
- the ignores are written with their reason: SIM105 repo-wide (try/except/pass
  is the deliberate shape for best-effort Win32/Qt teardown and each site
  carries a comment), plus per-line suppressions for ctypes/comtypes protocol
  attributes, deliberately long-lived file handles and local-time filenames
- real findings the new rules surfaced and this fixes: the Import Audio elapsed
  counter measured wall-clock time and would jump on a DST or NTP change, three
  zip() calls now state the equal-length invariant they rely on, and the
  Nemotron chunk test bound its loop variables through closures
- everything else is mechanical: import order, outdated syntax for Python 3.12,
  raw strings for the intentional regexes in pytest.raises(match=...), stale
  noqa directives, comprehensions, and merged nested with/if statements

+++ src/stt_app/model_download_progress.py
@@ -28,7 +28,7 @@ class ModelDownloadProgress:
     @property
     def percent(self) -> int | None:
         fraction = self.fraction
-        return None if fraction is None else int(round(fraction * 100))
+        return None if fraction is None else round(fraction * 100)
 
 
 class ModelDownloadSpeedTracker:
@@ -135,7 +135,7 @@ def format_model_download_progress(
         progress_bar = ""
         if include_progress_bar and progress.fraction is not None:
             width = 18
-            filled = int(round(progress.fraction * width))
+            filled = round(progress.fraction * width)
             progress_bar = f" [{'#' * filled}{'.' * (width - filled)}]"
         detail = (
             f"Downloading '{progress.model_name}': approx. {progress.percent}%"
